Move templates to templates directory
