--- router/edu/index.html.ejs
@@ -1,38 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <title>UPchieve EDU Admin</title>
-</head>
-
-<body class="container p-5">
-    <h1>UPchieve EDU Admin</h1>
-    <a href="/edu/questions/new" class="btn btn-secondary d-inline" style="height: 35px;">new question</a>
-
-    <ul class="list-group list-group-flush mt-5">
-        <% adminPages.forEach(entry => { %>
-         <% if (Array.isArray(entry)) { %>
-           <ul class="list-group list-group-horizontal-md">
-            <% entry.forEach(e => { %>
-              <li class="list-group-item small border-0">
-                <a href="/edu/<%= e.path %>"><%= e.label %></a>
-              </li>
-            <% }) %>
-           </ul>
-         <% } else { %>
-           <li class="list-group-item font-weight-bold border-bottom-0">
-             <a href="/edu/<%= entry.path %>"><%= entry.label %></a>
-           </li>
-         <% } %>
-        <% }) %>
-    </ul>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-</body>
-
-</html>
